+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #CHANGES
 **2.0.3**
 * Added getter for hour and minute
+* Added getter for hour and minute values
 
 **2.0.2**
 * Fixed wrong date in renderDay when render month view